提交代码

+++ src/SwiftCode.BSS.Common/Helper/TokenModelJwt.cs
@@ -1,8 +1,21 @@
 ﻿namespace SwiftCode.BSS.Common.Helper;
 
+/// <summary>
+/// 令牌
+/// </summary>
 public class TokenModelJwt
 {
-    public long Uid { get; set; }
+    /// <summary>
+    /// Id
+    /// </summary>
+    public int Uid { get; set; }
+    /// <summary>
+    /// 角色
+    /// </summary>
     public string Role { get; set; }
+    /// <summary>
+    /// 职能
+    /// </summary>
     public string Work { get; set; }
+
 }
